+++ cpu/mmu.h
@@ -51,7 +51,7 @@ public:
   void write8(uint32_t address, uint8_t value);
   uint32_t read8(uint32_t address);
   void write16(uint32_t address, uint16_t value);
-  uint32_t read16(uint32_t address);
+  uint16_t read16(uint32_t address);
   uint16_t read16s(uint32_t i);
 };
